+++ src/prefect/tasks/twitter/twitter.py
@@ -1,6 +1,7 @@
 from typing import Any
 
 import tweepy
+
 from prefect.client import Secret
 from prefect.core import Task
 from prefect.utilities.tasks import defaults_from_attrs