+++ tests/client/test_client.py
@@ -257,7 +257,7 @@ def test_client_deploy(monkeypatch):
         {"cloud.graphql": "http://my-cloud.foo", "cloud.auth_token": "secret_token"}
     ):
         client = Client()
-    flow = prefect.Flow(name="test")
+    flow = prefect.Flow(name="test", storage=prefect.environments.storage.Bytes())
     flow_id = client.deploy(flow, project_name="my-default-project")
     assert flow_id == "long-id"
 
@@ -272,15 +272,12 @@ def test_client_deploy_builds_flow(monkeypatch):
         {"cloud.graphql": "http://my-cloud.foo", "cloud.auth_token": "secret_token"}
     ):
         client = Client()
-    flow = prefect.Flow(name="test")
+    flow = prefect.Flow(name="test", storage=prefect.environments.storage.Bytes())
     flow_id = client.deploy(flow, project_name="my-default-project")
 
     ## extract POST info
     variables = json.loads(post.call_args[1]["json"]["variables"])
-    assert (
-        variables["input"]["serializedFlow"]["environment"]["serialized_flow"]
-        is not None
-    )
+    assert variables["input"]["serializedFlow"]["storage"] is not None
 
 
 def test_client_deploy_optionally_avoids_building_flow(monkeypatch):
@@ -298,9 +295,7 @@ def test_client_deploy_optionally_avoids_building_flow(monkeypatch):
 
     ## extract POST info
     variables = json.loads(post.call_args[1]["json"]["variables"])
-    assert (
-        variables["input"]["serializedFlow"]["environment"]["serialized_flow"] is None
-    )
+    assert variables["input"]["serializedFlow"]["storage"] is None
 
 
 def test_client_deploy_with_bad_proj_name(monkeypatch):